extract math utilities
